split search box
