+++ public/metadata.json
@@ -1,11 +1,11 @@
 {
   "projects": 482,
   "samples": 168464,
-  "phyla": 212,
-  "classes": 60,
-  "countries": 65,
+  "phyla": 60,
+  "classes": 212,
   "regions": 7,
-  "date": "2023-07-26",
-  "url": "https://doi.org/10.5281/zenodo.8186994",
-  "version": "0.1.0"
+  "countries": 65,
+  "date": "2023-08-08",
+  "url": "https://doi.org/10.5281/zenodo.8226567",
+  "version": "0.2.0"
 }
