+++ src/App.css
@@ -88,6 +88,12 @@ h2 {
   text-align: center;
 }
 
+h3 {
+  font-weight: 400;
+  font-size: 1.3rem;
+  text-align: center;
+}
+
 b,
 strong {
   font-weight: 600;
