+++ src/sections/Overview.tsx
@@ -49,7 +49,9 @@ const Overview = () => {
   ];
 
   return (
-    <>
+    <section>
+      <h2>Overview</h2>
+
       <p>
         Our dataset includes over {samplesRound.toLocaleString()} samples of
         publicly available 16S rRNA amplicon sequencing data, all processed
@@ -70,7 +72,7 @@ const Overview = () => {
           );
         })}
       </div>
-    </>
+    </section>
   );
 };
