+++ src/sections/Search.module.css
@@ -4,3 +4,8 @@
   justify-content: center;
   gap: 20px;
 }
+
+.table :where(th, td):not(:first-child) {
+  width: 100px;
+  text-align: center;
+}
